Separate chemical details popup from view chemicals screen
